initial refactorings
