+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/io/Encoding.kt
@@ -109,3 +109,17 @@ sealed class ListSerializerTemplate<ValueT>(
 
 object CertificateChainBase64UrlSerializer: ListSerializerTemplate<X509Certificate>(
     using = X509CertificateBase64UrlSerializer)
+
+/**
+ * Drops bytes at the start, or adds zero bytes at the start, until the [size] is reached
+ */
+fun ByteArray.ensureSize(size: Int): ByteArray = (this.size - size).let { toDrop ->
+    when {
+        toDrop > 0 -> this.copyOfRange(toDrop, this.size)
+        toDrop < 0 -> ByteArray(-toDrop) + this
+        else -> this
+    }
+}
+
+@Suppress("NOTHING_TO_INLINE")
+inline fun ByteArray.ensureSize(size: UInt) = ensureSize(size.toInt())
