+++ indispensable/src/jvmTest/kotlin/at/asitplus/signum/indispensable/Asn1EncodingTest.kt
@@ -1,15 +1,16 @@
 package at.asitplus.signum.indispensable
 
 import at.asitplus.signum.indispensable.asn1.*
-import at.asitplus.signum.indispensable.asn1.Asn1.BitString
-import at.asitplus.signum.indispensable.asn1.Asn1.Bool
-import at.asitplus.signum.indispensable.asn1.Asn1.Null
-import at.asitplus.signum.indispensable.asn1.Asn1.OctetString
-import at.asitplus.signum.indispensable.asn1.Asn1.OctetStringEncapsulating
-import at.asitplus.signum.indispensable.asn1.Asn1.PrintableString
-import at.asitplus.signum.indispensable.asn1.Asn1.ExplicitlyTagged
-import at.asitplus.signum.indispensable.asn1.Asn1.UtcTime
-import at.asitplus.signum.indispensable.asn1.Asn1.Utf8String
+import at.asitplus.signum.indispensable.asn1.encoding.Asn1.BitString
+import at.asitplus.signum.indispensable.asn1.encoding.Asn1.Bool
+import at.asitplus.signum.indispensable.asn1.encoding.Asn1.Null
+import at.asitplus.signum.indispensable.asn1.encoding.Asn1.OctetString
+import at.asitplus.signum.indispensable.asn1.encoding.Asn1.OctetStringEncapsulating
+import at.asitplus.signum.indispensable.asn1.encoding.Asn1.PrintableString
+import at.asitplus.signum.indispensable.asn1.encoding.Asn1.ExplicitlyTagged
+import at.asitplus.signum.indispensable.asn1.encoding.Asn1.UtcTime
+import at.asitplus.signum.indispensable.asn1.encoding.Asn1.Utf8String
+import at.asitplus.signum.indispensable.asn1.encoding.*
 import at.asitplus.signum.indispensable.io.BitSet
 import com.ionspin.kotlin.bignum.integer.BigInteger
 import com.ionspin.kotlin.bignum.integer.base63.toJavaBigInteger
@@ -77,7 +78,7 @@ class Asn1EncodingTest : FreeSpec({
                     }
                 }.derEncoded
             )
-            +ExplicitlyTagged(9u) { +Clock.System.now().encodeToAsn1UtcTime() }
+            +ExplicitlyTagged(9u) { +Clock.System.now().encodeToAsn1UtcTimePrimitive() }
             +OctetString(byteArrayOf(17, -43, 23, -12, 8, 65, 90))
             +Bool(false)
             +Bool(true)
@@ -92,7 +93,7 @@ class Asn1EncodingTest : FreeSpec({
             val bytes = (it).toTwosComplementByteArray()
 
             val fromBC = ASN1Integer(it).encoded
-            val long = Long.decodeFromDerValue(bytes)
+            val long = Long.decodeFromAsn1ContentBytes(bytes)
 
             val encoded = Asn1Primitive(Asn1Element.Tag.INT, bytes).derEncoded
             encoded shouldBe fromBC
