+++ indispensable/src/jvmTest/kotlin/at/asitplus/signum/indispensable/Asn1IntegerTest.kt
@@ -1,6 +1,9 @@
 package at.asitplus.signum.indispensable
 
 import at.asitplus.signum.indispensable.asn1.*
+import at.asitplus.signum.indispensable.asn1.encoding.encodeToAsn1Primitive
+import at.asitplus.signum.indispensable.asn1.encoding.parse
+import at.asitplus.signum.indispensable.asn1.encoding.readBigInteger
 import com.ionspin.kotlin.bignum.integer.BigInteger
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
@@ -8,12 +11,12 @@ import io.kotest.matchers.shouldBe
 class Asn1IntegerTest : FreeSpec({
     "Encoding: Negative" {
         val result =
-            BigInteger(-20).encodeToTlv()
+            BigInteger(-20).encodeToAsn1Primitive()
         result.toDerHexString() shouldBe "02 01 EC".replace(" ", "")
     }
     "Encoding: Large Positive" {
         val result =
-            BigInteger(0xEC).encodeToTlv()
+            BigInteger(0xEC).encodeToAsn1Primitive()
         result.toDerHexString() shouldBe "02 02 00 EC".replace(" ", "")
     }
     "Decoding: Negative" {
